refactor: post likes button refactor
